Move some gode to graph_interop.rs
