+++ blackjack_ui/src/application/serialization.rs
@@ -70,159 +70,71 @@ pub fn load(
     path: PathBuf,
     node_definitions: &NodeDefinitions,
 ) -> Result<(GraphEditorState, CustomGraphState)> {
-    // TODO: REVIEW: Should at least move some of this code to graph_interop.rs,
-    // where the other function is.
-
     let serialized = SerializedBjkGraph::load_from_file(&path)?;
     let (runtime, ui_data, id_idx_mappings) = serialized.to_runtime()?;
-    if let Some(ui_data) = ui_data {
-        // Create the graph and the id mappings
-        let mut graph = Graph::new();
-        let mut mapping = NodeMapping::new();
-        let BjkGraph { nodes: bjk_nodes } = runtime.graph;
 
-        // Fill in the nodes in a first pass
-        for (bjk_node_id, bjk_node) in &bjk_nodes {
-            let new_id = graph.add_node(
-                if let Some(node_def) = node_definitions.node_def(&bjk_node.op_name) {
-                    node_def.label.clone()
-                } else {
-                    "⚠ Unknown".into()
-                },
-                NodeData {
-                    op_name: bjk_node.op_name.clone(),
-                },
-                |_, _| { /* Params added later */ },
-            );
-            mapping.insert(new_id, bjk_node_id);
-        }
-
-        // Then, define inputs / outputs in a second pass.
-        for (bjk_node_id, bjk_node) in &bjk_nodes {
-            for bjk_input in &bjk_node.inputs {
-                graph.add_input_param(
-                    mapping[bjk_node_id],
-                    bjk_input.name.clone(),
-                    DataTypeUi(bjk_input.data_type),
-                    {
-                        let external_parameters = &runtime.external_parameters;
-                        let get_runtime_val = || -> BlackjackValue {
-                            // Try to get the value from the external parameters
-                            if let Some(ext) = external_parameters {
-                                let param = &ExternalParameter {
-                                    node_id: bjk_node_id,
-                                    param_name: bjk_input.name.clone(),
-                                };
-                                if let Some(val) = ext.0.get(&param) {
-                                    return val.clone();
-                                }
-                            }
-                            // Otherwise, try to get it from the node definition's default value
-                            if let Some(node_def) = node_definitions.node_def(&bjk_node.op_name) {
-                                if let Some(def) = node_def
-                                    .inputs
-                                    .iter()
-                                    .find(|input| input.name == bjk_input.name)
-                                {
-                                    return def.default_value();
-                                }
-                            }
-                            // If all else fails, return the default for the datatype.
-                            bjk_input.data_type.default_value()
-                        };
-                        ValueTypeUi(get_runtime_val())
-                    },
-                    data_type_to_input_param_kind(bjk_input.data_type),
-                    default_shown_inline(),
-                );
-            }
-
-            for bjk_output in &bjk_node.outputs {
-                graph.add_output_param(
-                    mapping[bjk_node_id],
-                    bjk_output.name.clone(),
-                    DataTypeUi(bjk_output.data_type),
-                );
-            }
-        }
-
-        // Finally, define connections in a third pass.
-        for (bjk_node_id, bjk_node) in &bjk_nodes {
-            for bjk_input in &bjk_node.inputs {
-                match &bjk_input.kind {
-                    DependencyKind::Connection { node, param_name } => {
-                        let out_node_id = mapping[*node];
-                        let out_id = graph[out_node_id]
-                            .get_output(&param_name)
-                            .expect("Param should exist, we just added it.");
-
-                        let in_node_id = mapping[bjk_node_id];
-                        let in_id = graph[in_node_id]
-                            .get_input(&bjk_input.name)
-                            .expect("Param should exist, we just added it.");
-
-                        graph.add_connection(out_id, in_id);
-                    }
-                    DependencyKind::External { .. } => {}
-                }
-            }
-        }
-
-        let idx_to_node_id = |idx| mapping[id_idx_mappings.get_id(idx).expect("Should exist")];
-
-        let node_order = ui_data
-            .node_order
-            .iter()
-            .map(|idx| idx_to_node_id(*idx))
-            .collect();
-
-        let node_positions = ui_data
-            .node_positions
-            .iter()
-            .enumerate()
-            .map(|(idx, pos)| (idx_to_node_id(idx), egui::pos2(pos.x, pos.y)))
-            .collect();
-
-        let mut promoted_params = HashMap::default();
-        for (bjk_node_id, bjk_node) in &bjk_nodes {
-            for bjk_input in &bjk_node.inputs {
-                if let DependencyKind::External {
-                    promoted: Some(promoted),
-                } = &bjk_input.kind
-                {
-                    let input_id = graph[mapping[bjk_node_id]]
-                        .get_input(&bjk_input.name)
-                        .expect("Should exist");
-                    promoted_params.insert(input_id, promoted.clone());
-                }
-            }
-        }
-
-        let editor_state = GraphEditorState {
-            graph,
-            node_order,
-            connection_in_progress: None,
-            selected_node: None,
-            node_positions,
-            node_finder: None,
-            pan_zoom: PanZoom {
-                pan: egui::vec2(ui_data.pan.x, ui_data.pan.y),
-                zoom: ui_data.zoom,
-            },
-            _user_state: std::marker::PhantomData,
-        };
-        let custom_state = CustomGraphState {
-            run_side_effect: None,
-            active_node: ui_data.active_node.map(idx_to_node_id),
-            node_definitions: node_definitions.share(),
-            promoted_params,
-        };
-
-        Ok((editor_state, custom_state))
-    } else {
+    if ui_data.is_none() {
         bail!(
             "The file at {} doesn't have UI information. Cannot load.",
             path.to_string_lossy()
         )
     }
+    let ui_data = ui_data.unwrap();
+
+    let (graph, mapping) = graph_interop::blackjack_graph_to_ui_graph(
+        &runtime.graph,
+        &runtime.external_parameters,
+        node_definitions,
+    )?;
+    let idx_to_node_id = |idx| mapping[id_idx_mappings.get_id(idx).expect("Should exist")];
+
+    let node_order = ui_data
+        .node_order
+        .iter()
+        .map(|idx| idx_to_node_id(*idx))
+        .collect();
+
+    let node_positions = ui_data
+        .node_positions
+        .iter()
+        .enumerate()
+        .map(|(idx, pos)| (idx_to_node_id(idx), egui::pos2(pos.x, pos.y)))
+        .collect();
+
+    let mut promoted_params = HashMap::default();
+    for (bjk_node_id, bjk_node) in &runtime.graph.nodes {
+        for bjk_input in &bjk_node.inputs {
+            if let DependencyKind::External {
+                promoted: Some(promoted),
+            } = &bjk_input.kind
+            {
+                let input_id = graph[mapping[bjk_node_id]]
+                    .get_input(&bjk_input.name)
+                    .expect("Should exist");
+                promoted_params.insert(input_id, promoted.clone());
+            }
+        }
+    }
+
+    let editor_state = GraphEditorState {
+        graph,
+        node_order,
+        connection_in_progress: None,
+        selected_node: None,
+        node_positions,
+        node_finder: None,
+        pan_zoom: PanZoom {
+            pan: egui::vec2(ui_data.pan.x, ui_data.pan.y),
+            zoom: ui_data.zoom,
+        },
+        _user_state: std::marker::PhantomData,
+    };
+    let custom_state = CustomGraphState {
+        run_side_effect: None,
+        active_node: ui_data.active_node.map(idx_to_node_id),
+        node_definitions: node_definitions.share(),
+        promoted_params,
+    };
+
+    Ok((editor_state, custom_state))
 }
