+++ blackjack_ui/src/graph/graph_interop.rs
@@ -6,11 +6,11 @@
 
 use std::ops::Index;
 
-use super::node_graph::{CustomGraphState, Graph};
+use super::node_graph::{CustomGraphState, Graph, NodeData, DataTypeUi, ValueTypeUi, data_type_to_input_param_kind, default_shown_inline};
 
 use crate::prelude::*;
 use blackjack_engine::{
-    graph::{BjkGraph, BjkNodeId, DependencyKind, NodeDefinitions},
+    graph::{BjkGraph, BjkNodeId, DependencyKind, NodeDefinitions, BlackjackValue},
     graph_interpreter::{ExternalParameter, ExternalParameterValues},
 };
 use egui_node_graph::{InputId, NodeId, OutputId};
@@ -29,7 +29,6 @@ impl NodeMapping {
         self.0.insert(node_id, bjk_node_id);
         self.1.insert(bjk_node_id, node_id);
     }
-
 }
 impl Index<NodeId> for NodeMapping {
     type Output = BjkNodeId;
@@ -90,6 +89,105 @@ pub fn ui_graph_to_blackjack_graph(
     }
 
     Ok((bjk_graph, NodeMapping(mapping, rev_mapping)))
+}
+
+pub fn blackjack_graph_to_ui_graph(
+    bjk_graph: &BjkGraph,
+    external_parameters: &Option<ExternalParameterValues>,
+    node_definitions: &NodeDefinitions,
+) -> Result<(Graph, NodeMapping)> {
+    // Create the graph and the id mappings
+    let mut graph = Graph::new();
+    let mut mapping = NodeMapping::new();
+    let BjkGraph { nodes: bjk_nodes } = bjk_graph;
+
+    // Fill in the nodes in a first pass
+    for (bjk_node_id, bjk_node) in bjk_nodes {
+        let new_id = graph.add_node(
+            if let Some(node_def) = node_definitions.node_def(&bjk_node.op_name) {
+                node_def.label.clone()
+            } else {
+                "⚠ Unknown".into()
+            },
+            NodeData {
+                op_name: bjk_node.op_name.clone(),
+            },
+            |_, _| { /* Params added later */ },
+        );
+        mapping.insert(new_id, bjk_node_id);
+    }
+
+    // Then, define inputs / outputs in a second pass.
+    for (bjk_node_id, bjk_node) in bjk_nodes {
+        for bjk_input in &bjk_node.inputs {
+            graph.add_input_param(
+                mapping[bjk_node_id],
+                bjk_input.name.clone(),
+                DataTypeUi(bjk_input.data_type),
+                {
+                    let get_runtime_val = || -> BlackjackValue {
+                        // Try to get the value from the external parameters
+                        if let Some(ext) = external_parameters {
+                            let param = &ExternalParameter {
+                                node_id: bjk_node_id,
+                                param_name: bjk_input.name.clone(),
+                            };
+                            if let Some(val) = ext.0.get(&param) {
+                                return val.clone();
+                            }
+                        }
+                        // Otherwise, try to get it from the node definition's default value
+                        if let Some(node_def) = node_definitions.node_def(&bjk_node.op_name) {
+                            if let Some(def) = node_def
+                                .inputs
+                                .iter()
+                                .find(|input| input.name == bjk_input.name)
+                            {
+                                return def.default_value();
+                            }
+                        }
+                        // If all else fails, return the default for the datatype.
+                        bjk_input.data_type.default_value()
+                    };
+                    ValueTypeUi(get_runtime_val())
+                },
+                data_type_to_input_param_kind(bjk_input.data_type),
+                default_shown_inline(),
+            );
+        }
+
+        for bjk_output in &bjk_node.outputs {
+            graph.add_output_param(
+                mapping[bjk_node_id],
+                bjk_output.name.clone(),
+                DataTypeUi(bjk_output.data_type),
+            );
+        }
+    }
+
+    // Finally, define connections in a third pass.
+    for (bjk_node_id, bjk_node) in bjk_nodes {
+        for bjk_input in &bjk_node.inputs {
+            match &bjk_input.kind {
+                DependencyKind::Connection { node, param_name } => {
+                    let out_node_id = mapping[*node];
+                    let out_id = graph[out_node_id]
+                        .get_output(&param_name)
+                        .expect("Param should exist, we just added it.");
+
+                    let in_node_id = mapping[bjk_node_id];
+                    let in_id = graph[in_node_id]
+                        .get_input(&bjk_input.name)
+                        .expect("Param should exist, we just added it.");
+
+                    graph.add_connection(out_id, in_id);
+                }
+                DependencyKind::External { .. } => {}
+            }
+        }
+    }
+
+    Ok((graph, mapping))
 }
 
 pub fn extract_graph_params(
